@@ -2,12 +2,13 @@
 
 from __future__ import annotations
 
+import types
+from collections.abc import Callable
 from dataclasses import dataclass
 from datetime import date, datetime, timedelta
 from typing import (
     Annotated,
     Any,
-    Callable,
     Literal,
     Union,
     get_args,
@@ -246,7 +247,8 @@ def _extract_type_and_validators(
         annotation = base_type
 
     origin = get_origin(annotation)
-    if origin is Union:
+    # Handle both Union[int, str] and int | str (PEP 604)
+    if origin is Union or isinstance(annotation, types.UnionType):
         args = get_args(annotation)
         # Check if None is in the union
         if type(None) in args:
